Move OverviewSettings to a .vue file. (#3471)
