+++ src/components/overview/OverviewSettings.vue
@@ -1,0 +1,27 @@
+<template>
+  <div class="players-overview-settings">
+    <div class="setting-button" v-on:click.prevent="toggleTagsView()">{{ getTagToggleLabel() }}</div>
+    <div class="setting-label">[ toggle tags view ]</div>
+  </div>
+</template>
+
+<script lang="ts">
+
+import Vue from 'vue';
+import {Shared} from './Shared';
+
+export default Vue.extend({
+  name: 'OverviewSettings',
+  methods: {
+    toggleTagsView: function() {
+      (this.$root as any).setVisibilityState(
+        'tags_concise',
+        !(this.$root as any).getVisibilityState('tags_concise'),
+      );
+    },
+    getTagToggleLabel: function(): string {
+      return Shared.isTagsViewConcise(this.$root) ? 'full' : 'concise';
+    },
+  },
+});
+</script>
